Add elapsed time chart to custom layout feature
